fixed log2 and next_pow2 macros to return a u64 number

+++ src/ion/headers/macros/next_pow2.h
@@ -3,4 +3,4 @@
  *
  * If the *number* is already a power of `2`, this macro is idempotent. */
 #define next_pow2(number) \
-  (1 << log2(number))
+  ((u64) 1 << log2(number))
